basic mechanics of all proof references working, going to re-work the locatio/exit menu thing and remove one indirection
