+++ src/lib/maptrie/logical.mocha.js
@@ -232,317 +232,6 @@ describe("LogicalTopology tests", function () {
     expect(proof.length).to.be.greaterThan(0);
   });
 
-  it("Should prove start location", async function () {
-    // build and verify a proof stack showing that a specific location exits are bound to exit menu choice inputs
-    const topo = new LogicalTopology();
-    topo.extendJoins([{ joins: [0, 1], sides: [3, 1] }]); // rooms 0,1 sides EAST, WEST
-    topo.extendLocations([
-      { sides: [[], [], [], [0]], flags: {} },
-      { sides: [[], [0], [], []], flags: {} },
-    ]);
-    const trie = topo.encodeTrie();
-
-    // mint without publishing nft metadata
-    let r = await this.mintGame({ topology: topo, trie });
-    const arenaEvents = new EventParser(this.arena, ArenaEvent.fromParsedEvent);
-    const gid = getGameCreated(r, arenaEvents).gid;
-    let transactor = new Transactor(arenaEvents);
-
-    let key = leafHash(
-      topo.prepareLeaf(
-        new LeafObject({
-          type: ObjectType.ExitMenu,
-          leaf: topo.locationExitMenu(0),
-        })
-      )
-    );
-    let id = topo.exitMenuKeys[key];
-
-    let refMenuInput = topo.referenceProofInput(ObjectType.ExitMenu, id, {
-      side: 3,
-      exit: 0,
-    });
-    const refLocation = new LogicalRef(
-      LogicalRefType.Proof,
-      ObjectType.Location2,
-      0
-    );
-
-    let egressExit = new LeafObject({
-      type: LocationExit.ObjectType,
-      leaf: new LocationExit(refMenuInput, refLocation),
-    });
-    let prepared = topo.prepareLeaf(egressExit);
-    let egressChoices = [
-      {
-        typeId: ObjectType.Exit,
-        inputs: egressExit.leaf.inputs({
-          resolveValue: topo.resolveValueRef.bind(topo),
-        }),
-      },
-    ];
-    let egressProofs = [trie.getProof(prepared)];
-
-    // id = topo.exitKeys[leafHash(topo.prepareLeaf(egressExit))];
-    // let refExit = new LogicalRef(LogicalRefType.Proof, ObjectType.Exit, id);
-
-    const startArgs = [
-      {
-        rootLabel: this.minter.minter.initArgs.rootLabels[0],
-        choices: egressChoices,
-        proofs: egressProofs,
-        data: [msgpack.encode({ sides: [[], [], [], [0]] })],
-      },
-    ];
-
-    transactor
-      .method(
-        this.user1Arena.registerTrialist,
-        gid,
-        msgpack.encode({ nickname: "alice" })
-      )
-      .requireLogs("TranscriptRegistration(uint256,address,bytes)")
-      .method(this.guardianArena.startTranscript, gid, startArgs[0])
-      .requireLogs(
-        "TranscriptStarted(uint256)",
-        "TranscriptEntryChoices(uint256,address,uint256,(uint256,bytes32[][]),bytes)"
-      );
-
-    // TODO: XXX is this failing because we condition the inputs to uint8array rather than hex string ?
-    for await (const r of transactor.transact()) {
-      console.log(
-        Object.keys(r.events).map((name) => `${name}[${r.events[name].length}]`)
-      );
-    }
-  });
-
-  it("Should prove location exit", async function () {
-    // build and verify a proof stack showing that a specific location exits are bound to exit menu choice inputs
-    const topo = new LogicalTopology();
-    topo.extendJoins([{ joins: [0, 1], sides: [3, 1] }]); // rooms 0,1 sides EAST, WEST
-    topo.extendLocations([
-      { sides: [[], [], [], [0]], flags: {} },
-      { sides: [[], [0], [], []], flags: {} },
-    ]);
-    const trie = topo.encodeTrie();
-
-    // mint without publishing nft metadata
-    let r = await this.mintGame({ topology: topo, trie });
-
-    const arenaEvents = new EventParser(this.arena, ArenaEvent.fromParsedEvent);
-    const gid = getGameCreated(r, arenaEvents).gid;
-    let transactor = new Transactor(arenaEvents);
-
-    let exitMenuKey = leafHash(
-      topo.prepareLeaf(
-        new LeafObject({
-          type: ObjectType.ExitMenu,
-          leaf: topo.locationExitMenu(0),
-        })
-      )
-    );
-    let id = topo.exitMenuKeys[exitMenuKey];
-
-    let refMenuInput = topo.referenceProofInput(ObjectType.ExitMenu, id, {
-      side: 3,
-      exit: 0,
-    });
-    let refLocation = new LogicalRef(
-      LogicalRefType.Proof,
-      ObjectType.Location2,
-      0
-    );
-
-    let egressExit = new LeafObject({
-      type: LocationExit.ObjectType,
-      leaf: new LocationExit(refMenuInput, refLocation),
-    });
-    let preparedExit = topo.prepareLeaf(egressExit);
-    let egressChoices = [
-      {
-        typeId: ObjectType.Exit,
-        inputs: egressExit.leaf.inputs({
-          resolveValue: topo.resolveValueRef.bind(topo),
-        }),
-      },
-    ];
-    let egressProofs = [trie.getProof(preparedExit)];
-
-    let ingressExitMenuKey = leafHash(
-      topo.prepareLeaf(
-        new LeafObject({
-          type: ObjectType.ExitMenu,
-          leaf: topo.locationExitMenu(1),
-        })
-      )
-    );
-    id = topo.exitMenuKeys[ingressExitMenuKey];
-    refMenuInput = topo.referenceProofInput(ObjectType.ExitMenu, id, {
-      side: 1,
-      exit: 0,
-    });
-    refLocation = new LogicalRef(LogicalRefType.Proof, ObjectType.Location2, 1);
-    let ingressExit = new LeafObject({
-      type: LocationExit.ObjectType,
-      leaf: new LocationExit(refMenuInput, refLocation),
-    });
-    let preparedIngressExit = topo.prepareLeaf(ingressExit);
-    let ingressChoice = [
-      {
-        typeId: ObjectType.Exit,
-        inputs: ingressExit.leaf.inputs({
-          resolveValue: topo.resolveValueRef.bind(topo),
-        }),
-      },
-    ];
-    let ingressProof = [trie.getProof(preparedIngressExit)];
-
-    // id = topo.exitKeys[leafHash(topo.prepareLeaf(egressExit))];
-    // let refExit = new LogicalRef(LogicalRefType.Proof, ObjectType.Exit, id);
-
-    const startArgs = [
-      {
-        rootLabel: this.minter.minter.initArgs.rootLabels[0],
-        choices: egressChoices,
-        proofs: egressProofs,
-        data: [msgpack.encode({ sides: [[], [], [], [0]] })],
-      },
-    ];
-
-    transactor
-      .method(
-        this.user1Arena.registerTrialist,
-        gid,
-        msgpack.encode({ nickname: "alice" })
-      )
-      .requireLogs("TranscriptRegistration(uint256,address,bytes)")
-      .method(this.guardianArena.startTranscript, gid, startArgs[0])
-      .requireLogs(
-        "TranscriptStarted(uint256)",
-        "TranscriptEntryChoices(uint256,address,uint256,(uint256,bytes32[][]),bytes)"
-      )
-      .method(this.user1Arena.transcriptEntryCommit, gid, {
-        rootLabel: this.minter.minter.initArgs.rootLabels[0],
-        node: leafHash(preparedExit),
-        data: "0x0",
-      })
-      .requireLogs(
-        "TranscriptEntryCommitted(uint256,address,uint256,bytes32,bytes32,bytes)"
-      );
-
-    // TODO: get the participant commit working
-    for await (const r of transactor.transact()) {
-      console.log(
-        Object.keys(r.events).map((name) => `${name}[${r.events[name].length}]`)
-      );
-    }
-
-    // The guardian will know the current location id for all players. To
-    // confirm a player exit choice the guardian must prove that the public
-    // player choice selects an exit in the current location that is linked to
-    // an exit in the next.
-
-    // For maximum 'fog' the locations would be blinded uniquely for each
-    // participant, and the relations would be duplicated in the trie for all
-    // players.
-
-    let leaves = [];
-    let stack = [];
-
-    const stackProofInputRef = (istack, jinput) => {
-      const hi = ethers.BigNumber.from(istack);
-      const lo = ethers.BigNumber.from(jinput);
-      hi.shl(16);
-      // Note: just assume lo doesn't need masking off for test convenience
-      return hi.or(lo).toHexString();
-    };
-
-    // Let the player start location id be 0 (the first location added above)
-
-    let startId = 0;
-
-    // Let the player chose the first exit on the east side.
-    let choice = 0; // input 0 is [3, 0]
-
-    // Obtain a proof for the egress exit menu
-    let prepared = [ObjectType.ExitMenu, directPreimage([[3, 0]])]; // side 3 (east), exit 0
-    let proof = trie.getProof(prepared);
-    // let key = leafHash(prepared);
-
-    // Set STACK(0) to exitMenu referenced by L0
-    leaves.push({
-      typeId: ObjectType.ExitMenu,
-      inputs: [[zeroPad(hexlify(3), 32), zeroPad(hexlify(0), 32)]],
-    });
-    stack.push({
-      inputRefs: [],
-      proofRefs: [],
-      rootLabel: "dungeon-state",
-      proof,
-    });
-
-    // Obtain a proof linking the exit menu to L0
-    // [LOCATION, [[locationId], [REF(#EM)]]]
-
-    prepared = [ObjectType.Location2, directPreimage([[0], [0]])];
-    proof = trie.getProof(prepared);
-    // STACK(1) to the association of location 0 with exitMenu 0
-    leaves.push({
-      typeId: ObjectType.Location2,
-      inputs: [[zeroPad(hexlify(0), 32)], [zeroPad(hexlify(0), 32)]],
-    });
-    stack.push({
-      inputRefs: [],
-      proofRefs: [1], // mark input[1] as reference to proven leaf node
-      rootLabel: "dungeon-state",
-      proof,
-    });
-
-    // Obtain an exit proof linking the exit menu choice to a specific location exit
-    // [EXIT, [[REF(#S, i)], [REF(#L)]]]
-    prepared = [ObjectType.Exit, directPreimage([[0, 0], [1]])];
-    proof = trie.getProof(prepared);
-    // STACK(2) to the association of exit 0 with location 0 with exitMenu 0, choice 0
-    leaves.push({
-      typeId: ObjectType.Exit,
-      inputs: [[stackProofInputRef(0, 0)], [zeroPad(hexlify(0), 32)]],
-    });
-    stack.push({
-      inputRefs: [0],
-      proofRefs: [1],
-      rootLabel: "dungeon-state",
-      proof,
-    });
-
-    // now add the proofs for the destination location, the ingress exit and the the choices at that location.
-
-    // finally, reveal (and prove) the outcome and choices consequent from the player choice
-    const user1Address = await this.user1Arena.signer.getAddress();
-
-    transactor
-      .method(this.guardianArena.transcriptEntryResolve, gid, {
-        participant: user1Address,
-        outcome: 3, // Accepted
-        stack,
-        leaves,
-        data: "0x",
-        choiceLeafIndex: 0, // XXX: TODO add proofs for the egress / ingress link and the ingress / menu link
-      })
-      .requireLogs(
-        "TranscriptEntryChoices(uint256,address,uint256,bytes32[],bytes)",
-        "TranscriptEntryOutcome(uint256,address,uint256,address,bytes32,uint8,bytes32,bytes)"
-      );
-
-    // TODO: get the participant commit working
-    for await (const r of transactor.transact()) {
-      console.log(
-        Object.keys(r.events).map((name) => `${name}[${r.events[name].length}]`)
-      );
-    }
-
-    // exitMenuKey is the the leaf for exit 0 (the start position of the player)
-  });
-
   it("Should throw because join and location access disagree", function () {
     const topo = new LogicalTopology();
     topo.extendJoins([{ joins: [0, 1], sides: [3, 1] }]); // rooms 0,1 sides EAST, WEST
